Debugging
I have attempted to compile the code and been trying to fix errors it gives me in the compiler. Right now only merge sort seems mad.

+++ DataClass.cpp
@@ -16,7 +16,7 @@
 Data userInputInfo()
 {
     //local variables to get values
-    int resistance, powerRating; 
+    double resistance, powerRating; 
     string colorCode; 
     Data newResistor;
